+++ src/app/Sidebar/components/searchInput.ts
@@ -36,19 +36,19 @@ const searchInput = (
   <div id="subtitle" class="flex items-center gap-4">
     <p class="italic">${subtitle}</p>
     <div class="flex gap-3">
-      <div id="registry-npm" class="action-item ${
+      <div id="registry-npm" class="action-item registry-action-item ${
         npm > 0 ? "flex" : "hidden"
-      } w-fit items-center justify-center rounded py-1.5 px-2 bg-[--vscode-toolbar-hoverBackground]">
-        <div class="flex items-center justify-center gap-1">
+      } w-fit items-center justify-center rounded py-1.5 px-2 bg-[--vscode-toolbar-hoverBackground] hover:cursor-pointer hover:brightness-90">
+        <div class="registry-data flex items-center justify-center gap-1">
           <div class="size-2.5 rounded-full bg-red-400"></div>
           <p id="npm-dependencies" class="font-semibold italic">${npm}</p>
         </div>
         <div class="tooltip tooltip-registry-subtitle tooltip-registry-npm-subtitle">Npm</div>
       </div>
-      <div id="registry-packagist" class="action-item ${
-        npm > 0 ? "flex" : "hidden"
-      } w-fit items-center justify-center rounded py-1.5 px-2 bg-[--vscode-toolbar-hoverBackground]">
-        <div class="flex items-center justify-center gap-1">
+      <div id="registry-packagist" class="action-item registry-action-item ${
+        packagist > 0 ? "flex" : "hidden"
+      } w-fit items-center justify-center rounded py-1.5 px-2 bg-[--vscode-toolbar-hoverBackground] hover:cursor-pointer hover:brightness-90">
+        <div class="registry-data flex items-center justify-center gap-1">
           <div class="size-2.5 rounded-full bg-yellow-400"></div>
           <p id="packagist-dependencies" class="font-semibold italic">${packagist}</p>
         </div>